+++ dvm/dvm/_.d
@@ -13,3 +13,4 @@ import dvm.dvm.CommandManager;
 import dvm.dvm.Exceptions;
 import dvm.dvm.Options;
 import dvm.dvm.ShellScript;
+import dvm.dvm.Version;